simplify code organisation
